+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The public API of this library consists of the public functions declared in
 file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java), and support
 for the Linux x64 and Darwin x64 platforms.
 
+## [3.6.0] - 2019-08-19
+### Added
+- `getPentagonIndexes` and `h3ToCenterChild` functions. (#49)
+### Changed
+- Updated the core library to v3.6.0. (#49)
+- Native implementations of `getRes0Indexes` and `getPentagonIndexes` changed to throw `OutOfMemoryError` if the output array size is too small. (#49)
+
 ## [3.5.0] - 2019-07-22
 ### Changed
 - Updated the core library to v3.5.0. (#47)